Move OmniSci backend to backends/ directory
