+++ ibis/backends/omniscidb/client.py
@@ -16,9 +16,9 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 from ibis.client import Database, DatabaseEntity, Query, SQLClient
-from ibis.omniscidb import ddl
-from ibis.omniscidb import dtypes as omniscidb_dtypes
-from ibis.omniscidb.compiler import OmniSciDBDialect, build_ast
+from . import ddl
+from . import dtypes as omniscidb_dtypes
+from .compiler import OmniSciDBDialect, build_ast
 from ibis.sql.compiler import DDL, DML
 from ibis.util import log
