+++ ibis/backends/omniscidb/tests/test_client.py
