+++ ibis/backends/omniscidb/tests/test_operations.py
